requirements.txt
streamlit==1.46.0
google-generativeai==0.8.5
gtts==2.5.3
nltk==3.9.1
opencv-python==4.11.0.86
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
psycopg2-binary==2.9.10
pyaudio==0.2.14
pyttsx3==2.71
speechrecognition==3.12.0
sqlalchemy==2.0.41
textblob==0.18.0.post0
vadersentiment==3.3.2
python-dotenv==1.0.0